# Close #1979: Create API for upcoming and previous events on the organizer profile screen