--- a/config/routes.rb
+++ b/config/routes.rb
@@ -441,6 +441,7 @@ Spree::Core::Engine.add_routes do
         resources :vendors do
           resources :nearby_places, only: %i[index]
           resources :vendor_photos
+          resources :events
         end
         resource :homepage_data, only: [:show]
         resources :homepage_sections, only: [:index]