--- a/db/migrate/20240926044328_add_vendor_to_spree_taxon.rb
+++ b/db/migrate/20240926044328_add_vendor_to_spree_taxon.rb
@@ -1,0 +1,5 @@
+class AddVendorToSpreeTaxon < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :spree_taxons, :vendor, index: true, foreign_key: { to_table: :spree_vendors }, if_not_exists: true
+  end
+end