### Before

```
2024-12-03 09:22:09.155928 D (2.676ms) ActiveRecord -- Spree::Taxon Count -- { :sql => "SELECT COUNT(*) FROM (SELECT 1 AS one FROM \"spree_taxons\" WHERE \"spree_taxons\".\"vendor_id\" = $1 AND (to_date >= '2024-12-03 02:22:09.016643') ORDER BY \"spree_taxons\".\"from_date\" ASC LIMIT $2 OFFSET $3) subquery_for_count", :binds => { :vendor_id => 9, :limit => 25, :offset => 0 }, :allocations => 549, :cached => nil }
2024-12-03 09:22:09.161216 D (3.982ms) ActiveRecord -- Spree::Taxon Load -- { :sql => "SELECT \"spree_taxons\".* FROM \"spree_taxons\" WHERE \"spree_taxons\".\"vendor_id\" = $1 AND (to_date >= '2024-12-03 02:22:09.016643') ORDER BY \"spree_taxons\".\"from_date\" ASC LIMIT $2 OFFSET $3", :binds => { :vendor_id => 9, :limit => 25, :offset => 0 }, :allocations => 741, :cached => nil }
2024-12-03 09:22:09.189224 D (2.462ms) ActiveRecord -- Spree::Taxon Load -- { :sql => "SELECT \"spree_taxons\".* FROM \"spree_taxons\" WHERE \"spree_taxons\".\"lft\" <= 146 AND \"spree_taxons\".\"rgt\" >= 149 AND (\"spree_taxons\".\"id\" != 412) ORDER BY lft", :allocations => 232, :cached => nil }
2024-12-03 09:22:09.199594 D (1.448ms) ActiveRecord -- Spree::Taxon Pluck -- { :sql => "SELECT \"spree_taxons\".\"id\" FROM \"spree_taxons\" WHERE \"spree_taxons\".\"parent_id\" = $1 ORDER BY \"spree_taxons\".\"lft\" ASC", :binds => { :parent_id => 412 }, :allocations => 216, :cached => nil }
2024-12-03 09:22:09.223478 D (3.590ms) ActiveRecord -- Spree::TaxonImage Load -- { :sql => "SELECT \"spree_assets\".* FROM \"spree_assets\" WHERE \"spree_assets\".\"type\" = $1 AND \"spree_assets\".\"viewable_id\" = $2 AND \"spree_assets\".\"viewable_type\" = $3 LIMIT $4", :binds => { :type => "Spree::TaxonImage", :viewable_id => 412, :viewable_type => "Spree::Taxon", :limit => 1 }, :allocations => 233, :cached => nil }
2024-12-03 09:22:09.259730 D (2.987ms) ActiveRecord -- Spree::Vendor Pluck -- { :sql => "SELECT \"spree_vendors\".\"id\" FROM \"spree_vendors\" INNER JOIN \"cm_taxon_vendors\" ON \"spree_vendors\".\"id\" = \"cm_taxon_vendors\".\"vendor_id\" WHERE \"spree_vendors\".\"deleted_at\" IS NULL AND \"cm_taxon_vendors\".\"taxon_id\" = $1", :binds => { :taxon_id => 412 }, :allocations => 217, :cached => nil }
2024-12-03 09:22:09.287455 D (2.306ms) ActiveRecord -- Spree::Product Pluck -- { :sql => "SELECT \"spree_products\".\"id\" FROM \"spree_products\" INNER JOIN \"spree_products_taxons\" ON \"spree_products\".\"id\" = \"spree_products_taxons\".\"product_id\" WHERE \"spree_products\".\"deleted_at\" IS NULL AND \"spree_products_taxons\".\"taxon_id\" = $1 AND \"spree_products_taxons\".\"visible\" = $2 ORDER BY \"spree_products_taxons\".\"position\" ASC", :binds => { :taxon_id => 412, :visible => true }, :allocations => 217, :cached => nil }
```

### After Resolve N+1

```
2024-12-03 09:25:02.131561 D (1.877ms) ActiveRecord -- Spree::Taxon Count -- { :sql => "SELECT COUNT(*) FROM (SELECT 1 AS one FROM \"spree_taxons\" WHERE \"spree_taxons\".\"vendor_id\" = $1 AND (to_date >= '2024-12-03 02:25:01.991236') ORDER BY \"spree_taxons\".\"from_date\" ASC LIMIT $2 OFFSET $3) subquery_for_count", :binds => { :vendor_id => 9, :limit => 25, :offset => 0 }, :allocations => 576, :cached => nil }
2024-12-03 09:25:02.134417 D (2.049ms) ActiveRecord -- Spree::Taxon Load -- { :sql => "SELECT \"spree_taxons\".* FROM \"spree_taxons\" WHERE \"spree_taxons\".\"vendor_id\" = $1 AND (to_date >= '2024-12-03 02:25:01.991236') ORDER BY \"spree_taxons\".\"from_date\" ASC LIMIT $2 OFFSET $3", :binds => { :vendor_id => 9, :limit => 25, :offset => 0 }, :allocations => 714, :cached => nil }
2024-12-03 09:25:02.179799 D (1.690ms) ActiveRecord -- Spree::Vendor Load -- { :sql => "SELECT \"spree_vendors\".* FROM \"spree_vendors\" WHERE \"spree_vendors\".\"deleted_at\" IS NULL AND \"spree_vendors\".\"id\" = $1", :binds => { :id => 9 }, :allocations => 228, :cached => nil }
2024-12-03 09:25:02.194861 D (2.179ms) ActiveRecord -- Spree::Taxon Load -- { :sql => "SELECT \"spree_taxons\".* FROM \"spree_taxons\" WHERE \"spree_taxons\".\"lft\" <= 146 AND \"spree_taxons\".\"rgt\" >= 149 AND (\"spree_taxons\".\"id\" != 412) ORDER BY lft", :allocations => 520, :cached => nil }
2024-12-03 09:25:02.202104 D (1.359ms) ActiveRecord -- Spree::Taxon Pluck -- { :sql => "SELECT \"spree_taxons\".\"id\" FROM \"spree_taxons\" WHERE \"spree_taxons\".\"parent_id\" = $1 ORDER BY \"spree_taxons\".\"lft\" ASC", :binds => { :parent_id => 412 }, :allocations => 216, :cached => nil }
2024-12-03 09:25:02.221610 D (2.457ms) ActiveRecord -- Spree::TaxonImage Load -- { :sql => "SELECT \"spree_assets\".* FROM \"spree_assets\" WHERE \"spree_assets\".\"type\" = $1 AND \"spree_assets\".\"viewable_id\" = $2 AND \"spree_assets\".\"viewable_type\" = $3 LIMIT $4", :binds => { :type => "Spree::TaxonImage", :viewable_id => 412, :viewable_type => "Spree::Taxon", :limit => 1 }, :allocations => 233, :cached => nil }
2024-12-03 09:25:02.238408 D (2.974ms) ActiveRecord -- Spree::Vendor Pluck -- { :sql => "SELECT \"spree_vendors\".\"id\" FROM \"spree_vendors\" INNER JOIN \"cm_taxon_vendors\" ON \"spree_vendors\".\"id\" = \"cm_taxon_vendors\".\"vendor_id\" WHERE \"spree_vendors\".\"deleted_at\" IS NULL AND \"cm_taxon_vendors\".\"taxon_id\" = $1", :binds => { :taxon_id => 412 }, :allocations => 217, :cached => nil }
2024-12-03 09:25:02.268649 D (2.520ms) ActiveRecord -- Spree::Product Pluck -- { :sql => "SELECT \"spree_products\".\"id\" FROM \"spree_products\" INNER JOIN \"spree_products_taxons\" ON \"spree_products\".\"id\" = \"spree_products_taxons\".\"product_id\" WHERE \"spree_products\".\"deleted_at\" IS NULL AND \"spree_products_taxons\".\"taxon_id\" = $1 AND \"spree_products_taxons\".\"visible\" = $2 ORDER BY \"spree_products_taxons\".\"position\" ASC", :binds => { :taxon_id => 412, :visible => true }, :allocations => 217, :cached => nil }
```

### App: https://github.com/bookmebus/cm-market-app/pull/1566

- #### BaseUrl: {{BASE_URL}}/api/v2/storefront/vendors/9/events?section=all

``` json
{
    "data": [
        {
            "id": "394",
            "type": "taxon",
            "attributes": {
                "name": "The Q - Solo Concert",
                "pretty_name": "Events -> The Q - Solo Concert",
                "permalink": "events/ask-concert",
                "seo_title": "The Q - Solo Concert",
                "description": "<p>Immerse yourself in an exceptional musical experience at \"The Q Concert - Solo Concert\" featuring the acclaimed artist Aok Sokunkanha. This exclusive event promises an intimate and memorable performance in the heart of Phnom Penh.</p>\r\n<p><strong>Ticket Options:</strong></p>\r\n<ul>\r\n<li><strong>VIP:</strong> Enjoy unparalleled access with prime seating, entry to a dedicated VIP lounge, and a unique opportunity for a meet-and-greet with Aok Sokunkanha.</li>\r\n<li><strong>Premium:</strong> Opt for premium seating with superb views and access to exclusive event amenities.</li>\r\n<li><strong>Standard:</strong> Choose a standard ticket for excellent value and access to general seating.</li>\r\n</ul>\r\n<p>Explore the vibrant cultural scene in Cambodia and be part of this extraordinary evening of live music. Secure your tickets today and join us for a night of unforgettable entertainment.</p>",
                "meta_title": "The Q - Solo Concert",
                "meta_description": "Immerse yourself in an exceptional musical experience at \"The Q Concert - Solo Concert\" featuring the acclaimed artist Aok Sokunkanha. This exclusive event promises an intimate and memorable performance in the heart of Phnom Penh.\r\n\r\n",
                "meta_keywords": "",
                "left": 130,
                "right": 133,
                "position": 0,
                "depth": 1,
                "updated_at": "2024-11-06T15:03:26.176+07:00",
                "public_metadata": {},
                "is_root": false,
                "is_child": true,
                "is_leaf": false,
                "custom_redirect_url": "bookmeplus:/t/events/ask-concert",
                "kind": "event",
                "subtitle": "#asksoloconcert",
                "from_date": "2024-09-02T12:00:00.000+07:00",
                "to_date": "2024-10-16T12:00:00.000+07:00",
                "background_color": "#3B7A7A",
                "foreground_color": "#FFFFFF",
                "show_badge_status": true,
                "vendor_id": 9
            },
            "relationships": {
                "parent": {
                    "data": {
                        "id": "17",
                        "type": "taxon"
                    }
                },
                "taxonomy": {
                    "data": {
                        "id": "5",
                        "type": "taxonomy"
                    }
                },
                "children": {
                    "data": [
                        {
                            "id": "393",
                            "type": "taxon"
                        }
                    ]
                },
                "image": {
                    "data": null
                },
                "vendors": {
                    "data": []
                },
                "visible_products": {
                    "data": []
                },
                "category_icon": {
                    "data": {
                        "id": "1391",
                        "type": "asset"
                    }
                },
                "app_banner": {
                    "data": {
                        "id": "1450",
                        "type": "asset"
                    }
                },
                "web_banner": {
                    "data": {
                        "id": "1449",
                        "type": "asset"
                    }
                },
                "home_banner": {
                    "data": {
                        "id": "1451",
                        "type": "asset"
                    }
                }
            }
        },
        {
            "id": "412",
            "type": "taxon",
            "attributes": {
                "name": "Run With Sai",
                "pretty_name": "Events -> Run With Sai",
                "permalink": "events/run-with-sai-1",
                "seo_title": "Run With Sai",
                "description": "<p>Run with Sai: 400,000 steps to save children with cancer&rdquo;,&nbsp;<strong>a charity run initiated by Cambodian singer and performer Uon Pakthom, better known as Sai</strong>, officially kicked off on October 22</p>\r\n<p>🗓 Date: 22th-26th October 2024</p>\r\n<p>📍 Location: <a href=\"https://www.google.com/maps/place/National+University+of+Battambang+%2F+%E1%9E%9F%E1%9E%B6%E1%9E%80%E1%9E%9B%E1%9E%9C%E1%9E%B7%E1%9E%91%E1%9F%92%E1%9E%99%E1%9E%B6%E1%9E%9B%E1%9F%90%E1%9E%99%E2%80%8B%E1%9E%87%E1%9E%B6%E1%9E%8F%E1%9E%B7%E1%9E%94%E1%9E%B6%E1%9E%8F%E1%9F%8B%E1%9E%8A%E1%9F%86%E1%9E%94%E1%9E%84/@13.0862629,103.2197316,15z/data=!4m6!3m5!1s0x3105499fe765395f:0xafe041b8dd5dbb8a!8m2!3d13.0862629!4d103.2197316!16s%2Fm%2F0cpbht2?entry=ttu&amp;g_ep=EgoyMDI0MDgyMy4wIKXMDSoASAFQAw%3D%3D?target=_blank\">National University of Battambang</a></p>\r\n<p>⌚️ Time: From 7:00AM - 4:30PM</p>\r\n<p>🎫 Entry Fee: Free</p>",
                "meta_title": "",
                "meta_description": "",
                "meta_keywords": "",
                "left": 142,
                "right": 145,
                "position": 0,
                "depth": 1,
                "updated_at": "2024-11-06T14:42:43.771+07:00",
                "public_metadata": {},
                "is_root": false,
                "is_child": true,
                "is_leaf": false,
                "custom_redirect_url": "bookmeplus://www.bookme.plus/t/events/run-with-sai-1",
                "kind": "event",
                "subtitle": "Preah Vihear",
                "from_date": "2024-10-01T12:00:00.000+07:00",
                "to_date": "2024-10-31T12:00:00.000+07:00",
                "background_color": "#6666ff",
                "foreground_color": "",
                "show_badge_status": false,
                "vendor_id": 9
            },
            "relationships": {
                "parent": {
                    "data": {
                        "id": "17",
                        "type": "taxon"
                    }
                },
                "taxonomy": {
                    "data": {
                        "id": "5",
                        "type": "taxonomy"
                    }
                },
                "children": {
                    "data": [
                        {
                            "id": "413",
                            "type": "taxon"
                        }
                    ]
                },
                "image": {
                    "data": null
                },
                "vendors": {
                    "data": [
                        {
                            "id": "9",
                            "type": "vendor"
                        }
                    ]
                },
                "visible_products": {
                    "data": []
                },
                "category_icon": {
                    "data": {
                        "id": "1609",
                        "type": "asset"
                    }
                },
                "app_banner": {
                    "data": {
                        "id": "1620",
                        "type": "asset"
                    }
                },
                "web_banner": {
                    "data": {
                        "id": "1555",
                        "type": "asset"
                    }
                },
                "home_banner": {
                    "data": {
                        "id": "1556",
                        "type": "asset"
                    }
                }
            }
        },
        {
            "id": "395",
            "type": "taxon",
            "attributes": {
                "name": "The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2",
                "pretty_name": "Events -> The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2",
                "permalink": "events/the-q-solo-concert-2",
                "seo_title": "The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2 The Q - Solo Concert 2",
                "description": "<p>Immerse yourself in an exceptional musical experience at \"The Q Concert - Solo Concert\" featuring the acclaimed artist Aok Sokunkanha. This exclusive event promises an intimate and memorable performance in the heart of Phnom Penh.</p>\r\n<p><strong>Ticket Options:</strong></p>\r\n<ul>\r\n<li><strong>VIP:</strong> Enjoy unparalleled access with prime seating, entry to a dedicated VIP lounge, and a unique opportunity for a meet-and-greet with Aok Sokunkanha.</li>\r\n<li><strong>Premium:</strong> Opt for premium seating with superb views and access to exclusive event amenities.</li>\r\n<li><strong>Standard:</strong> Choose a standard ticket for excellent value and access to general seating.</li>\r\n</ul>\r\n<p>Explore the vibrant cultural scene in Cambodia and be part of this extraordinary evening of live music. Secure your tickets today and join us for a night of unforgettable entertainment.</p>",
                "meta_title": "",
                "meta_description": "Immerse yourself in an exceptional musical experience at \"The Q Concert - Solo Concert\" featuring the acclaimed artist Aok Sokunkanha. This exclusive event promises an intimate and memorable performance in the heart of Phnom Penh.",
                "meta_keywords": "",
                "left": 118,
                "right": 121,
                "position": 0,
                "depth": 1,
                "updated_at": "2024-11-06T15:02:39.418+07:00",
                "public_metadata": {},
                "is_root": false,
                "is_child": true,
                "is_leaf": false,
                "custom_redirect_url": "bookmeplus://www.bookme.plus/t/events/the-q-solo-concert-2",
                "kind": "event",
                "subtitle": " #theQsoloconcert #theQsoloconcert #theQsoloconcert #theQsoloconcert",
                "from_date": "2024-10-14T12:00:00.000+07:00",
                "to_date": "2024-10-31T12:00:00.000+07:00",
                "background_color": "#3B7A7A",
                "foreground_color": "#FFFFFF",
                "show_badge_status": false,
                "vendor_id": 9
            },
            "relationships": {
                "parent": {
                    "data": {
                        "id": "17",
                        "type": "taxon"
                    }
                },
                "taxonomy": {
                    "data": {
                        "id": "5",
                        "type": "taxonomy"
                    }
                },
                "children": {
                    "data": [
                        {
                            "id": "396",
                            "type": "taxon"
                        }
                    ]
                },
                "image": {
                    "data": null
                },
                "vendors": {
                    "data": [
                        {
                            "id": "145",
                            "type": "vendor"
                        }
                    ]
                },
                "visible_products": {
                    "data": []
                },
                "category_icon": {
                    "data": {
                        "id": "1461",
                        "type": "asset"
                    }
                },
                "app_banner": {
                    "data": {
                        "id": "1463",
                        "type": "asset"
                    }
                },
                "web_banner": {
                    "data": {
                        "id": "1462",
                        "type": "asset"
                    }
                },
                "home_banner": {
                    "data": {
                        "id": "1464",
                        "type": "asset"
                    }
                }
            }
        },
        {
            "id": "358",
            "type": "taxon",
            "attributes": {
                "name": "EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia",
                "pretty_name": "Events -> EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia",
                "permalink": "events/edm-cambodia",
                "seo_title": "EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia EDM Cambodia",
                "description": "",
                "meta_title": "",
                "meta_description": "",
                "meta_keywords": "",
                "left": 122,
                "right": 125,
                "position": 0,
                "depth": 1,
                "updated_at": "2024-11-06T14:43:54.669+07:00",
                "public_metadata": {},
                "is_root": false,
                "is_child": true,
                "is_leaf": false,
                "custom_redirect_url": "bookmeplus://www.bookme.plus/t/events/edm-cambodia",
                "kind": "event",
                "subtitle": "Electronic Dance Music",
…