@channainfo 2024-11-08
@Sreyleak-Deth it seems that you have the n+1 for the ActiveStorage for both the Blob and Attachement.